Move common test code to utils.c
