+++ c_bindings/Makefile
@@ -14,7 +14,7 @@ IINC       := -I include
 WARNINGS   := -Wall -Wextra -Wshadow -Werror
 
 
-C_BINDINGS_TESTS := $(wildcard tests/*.c)
+C_BINDINGS_TESTS := $(wildcard tests/test_*.c)
 C_BINDINGS_ELFS  := $(C_BINDINGS_TESTS:.c=.elf)
 
 all: $(C_BINDINGS_ELFS)
@@ -33,7 +33,7 @@ $(shell cargo rustc --manifest-path ../lib/Cargo.toml --lib --features c_binding
 
 
 %.elf: %.c $(LIB)
-	$(CC) $(CSTD) $(CFLAGS) $(IINC) $(WARNINGS) -o $@ $< -L ../target/$(BUILD_MODE) -Wl,-Bstatic -l crunch64 -Wl,-Bdynamic
+	$(CC) $(CSTD) $(CFLAGS) $(IINC) $(WARNINGS) -o $@ tests/utils.c $< -L ../target/$(BUILD_MODE) -Wl,-Bstatic -l crunch64 -Wl,-Bdynamic
 
 # Print target for debugging
 print-% : ; $(info $* is a $(flavor $*) variable set to [$($*)]) @true
